+++ src/plugins/storage-remote/rx-storage-remote.ts
@@ -31,14 +31,13 @@ import type {
     RxStorageRemoteInternals,
     RxStorageRemoteSettings
 } from './storage-remote-types';
-
-
+import { closeMessageChannel, getMessageChannel } from './message-channel-cache';
 
 
 export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any> {
     public readonly statics: RxStorageStatics;
     public readonly name: string = 'remote';
-    private requestIdSeed: string = randomCouchString(10);
+    private seed: string = randomCouchString(10);
     private lastRequestId: number = 0;
     constructor(
         public readonly settings: RxStorageRemoteSettings
@@ -48,7 +47,7 @@ export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any>
 
     public getRequestId() {
         const newId = this.lastRequestId++;
-        return this.requestIdSeed + '|' + newId;
+        return this.seed + '|' + newId;
     }
 
     async createStorageInstance<RxDocType>(
@@ -56,11 +55,29 @@ export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any>
     ): Promise<RxStorageInstanceRemote<RxDocType>> {
         const connectionId = 'c|' + this.getRequestId();
 
+        const cacheKeys: string[] = [
+            'mode-' + this.settings.mode
+        ];
+        switch (this.settings.mode) {
+            case 'collection':
+                cacheKeys.push('collection-' + params.collectionName);
+            // eslint-disable-next-line no-fallthrough
+            case 'database':
+                cacheKeys.push('database-' + params.databaseName);
+            // eslint-disable-next-line no-fallthrough
+            case 'storage':
+                cacheKeys.push('seed-' + this.seed);
+        }
+        const messageChannel = await getMessageChannel(
+            this.settings,
+            cacheKeys
+        );
+
         const requestId = this.getRequestId();
-        const waitForOkPromise = firstValueFrom(this.settings.messages$.pipe(
+        const waitForOkPromise = firstValueFrom(messageChannel.messages$.pipe(
             filter(msg => msg.answerTo === requestId)
         ));
-        this.settings.send({
+        messageChannel.send({
             connectionId,
             method: 'create',
             requestId,
@@ -71,6 +88,7 @@ export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any>
         if (waitForOkResult.error) {
             throw new Error('could not create instance ' + JSON.stringify(waitForOkResult.error));
         }
+
         return new RxStorageInstanceRemote(
             this,
             params.databaseName,
@@ -78,19 +96,21 @@ export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any>
             params.schema,
             {
                 params,
-                connectionId
+                connectionId,
+                messageChannel
             },
             params.options
         );
     }
 
     async customRequest<In, Out>(data: In): Promise<Out> {
+        const messageChannel = await this.settings.messageChannelCreator();
         const requestId = this.getRequestId();
         const connectionId = 'custom|request|' + requestId;
-        const waitForAnswerPromise = firstValueFrom(this.settings.messages$.pipe(
+        const waitForAnswerPromise = firstValueFrom(messageChannel.messages$.pipe(
             filter(msg => msg.answerTo === requestId)
         ));
-        this.settings.send({
+        messageChannel.send({
             connectionId,
             method: 'custom',
             requestId,
@@ -98,13 +118,16 @@ export class RxStorageRemote implements RxStorage<RxStorageRemoteInternals, any>
         });
         const response = await waitForAnswerPromise;
         if (response.error) {
+            await messageChannel.close();
             throw new Error('could not run customRequest(): ' + JSON.stringify({
                 data,
                 error: response.error
             }));
         } else {
+            await messageChannel.close();
             return response.return;
         }
+
     }
 }
 
@@ -124,7 +147,7 @@ export class RxStorageInstanceRemote<RxDocType> implements RxStorageInstance<RxD
         public readonly internals: RxStorageRemoteInternals,
         public readonly options: Readonly<any>
     ) {
-        this.messages$ = this.storage.settings.messages$.pipe(
+        this.messages$ = this.internals.messageChannel.messages$.pipe(
             filter(msg => msg.connectionId === this.internals.connectionId)
         );
         this.subs.push(
@@ -155,7 +178,7 @@ export class RxStorageInstanceRemote<RxDocType> implements RxStorageInstance<RxD
             method: methodName,
             params
         };
-        this.storage.settings.send(message);
+        this.internals.messageChannel.send(message);
         const response = await responsePromise;
         if (response.error) {
             throw new Error('could not requestRemote: ' + JSON.stringify({
@@ -209,10 +232,12 @@ export class RxStorageInstanceRemote<RxDocType> implements RxStorageInstance<RxD
         this.subs.forEach(sub => sub.unsubscribe());
         this.changes$.complete();
         await this.requestRemote('close', []);
+        await closeMessageChannel(this.internals.messageChannel);
     }
     async remove(): Promise<void> {
         this.closed = true;
         await this.requestRemote('remove', []);
+        await closeMessageChannel(this.internals.messageChannel);
     }
     conflictResultionTasks(): Observable<RxConflictResultionTask<RxDocType>> {
         return this.conflicts$;
@@ -223,5 +248,8 @@ export class RxStorageInstanceRemote<RxDocType> implements RxStorageInstance<RxD
 }
 
 export function getRxStorageRemote(settings: RxStorageRemoteSettings): RxStorageRemote {
-    return new RxStorageRemote(settings);
+    const withDefaults = Object.assign({
+        mode: 'storage'
+    }, settings);
+    return new RxStorageRemote(withDefaults);
 }