+++ src/plugins/utils/utils-map.ts
@@ -1,5 +1,14 @@
 
 
+
+export function getFromMapOrThrow<K, V>(map: Map<K, V> | WeakMap<any, V>, key: K): V {
+    const val = map.get(key);
+    if (typeof val === 'undefined') {
+        throw new Error('missing value from map ' + key);
+    }
+    return val;
+}
+
 export function getFromMapOrCreate<MapIndex, MapValue>(
     map: Map<MapIndex, MapValue> | WeakMap<any, MapValue>,
     index: MapIndex,