+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 <!-- CHANGELOG NEWEST -->
 - FIX `requestIdlePromise()` must run in a queue.
 - ADD Export ReplicationOptions type [#4606](https://github.com/pubkey/rxdb/pull/4606)
+- ADD `mode` option to remote storage
 <!-- ADD new changes here! -->
 
 <!-- /CHANGELOG NEWEST -->